feat: pg pool connections

+++ migrations/000001_create_metrics_table.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE metrics (
     id VARCHAR(255) NOT NULL,
     m_type VARCHAR(255) NOT NULL,
-    delta INTEGER,
+    delta BIGINT,
     value double precision,
     PRIMARY KEY (id, m_type)
 );
